// Merge -> revision 8507

+++ modules/autoupgrade/autoupgrade.php
@@ -26,21 +26,15 @@ class Autoupgrade extends Module
 	}
 	function install()
 	{
-		$autoupgradeCanWork = true;
-		if (!$autoupgradeCanWork)
-			return false;
 
 		$res = true;
 		// before adding AdminSelfUpgrade, we should remove AdminUpgrade
-		if (version_compare(_PS_VERSION_,'1.4.4.0','==') OR version_compare(_PS_VERSION_,'1.4.4.1','=='))
-		{
 			$idTab = Tab::getIdFromClassName('AdminUpgrade');
 
 			if ($idTab)
 			{
 			$tab = new Tab($idTab);
 				$res &= $tab->delete();
-		}
 		}
 		
 		$idTab = Tab::getIdFromClassName('AdminSelfUpgrade');
@@ -55,9 +49,7 @@ class Autoupgrade extends Module
 			$res &= $tab->save();
 		}
 		else
-		{
 			$tab = new Tab($idTab);
-		}
 			Configuration::updateValue('PS_AUTOUPDATE_MODULE_IDTAB',$tab->id);
 
 		$autoupgradeDir = _PS_ADMIN_DIR_.DIRECTORY_SEPARATOR.'autoupgrade';
@@ -68,6 +60,7 @@ class Autoupgrade extends Module
 		$path = dirname(__FILE__).'/';
 		
 		$res &= copy($path.'ajax-upgradetab.php',$autoupgradeDir . DIRECTORY_SEPARATOR . 'ajax-upgradetab.php');
+		$res &= copy($path.'logo.gif',_PS_ROOT_DIR_. DIRECTORY_SEPARATOR . 'img/t/AdminSelfUpgrade.gif');
 		
 		if (!$res 
 			OR !Tab::getIdFromClassName('AdminSelfUpgrade')
@@ -94,6 +87,7 @@ class Autoupgrade extends Module
 			$res = $tab->save();
 		}
 		
+		if (file_exists(_PS_ADMIN_DIR_.DIRECTORY_SEPARATOR.'autoupgrade'.DIRECTORY_SEPARATOR.'ajax-upgradetab.php'))
 		$res &= @unlink(_PS_ADMIN_DIR_.DIRECTORY_SEPARATOR.'autoupgrade'.DIRECTORY_SEPARATOR.'ajax-upgradetab.php');
 		if($res OR !parent::uninstall())
 			return false;
